// src/http/generated/lunajoy-server.schemas.ts
export type CreateDailyLogBodySocialInteractionsItemInteractionType = typeof CreateDailyLogBodySocialInteractionsItemInteractionType[keyof typeof CreateDailyLogBodySocialInteractionsItemInteractionType];


export const CreateDailyLogBodySocialInteractionsItemInteractionType = {
  FAMILY: 'FAMILY',
  FRIENDS: 'FRIENDS',
  COLLEAGUES: 'COLLEAGUES',
  STRANGERS: 'STRANGERS',
  WORK: 'WORK',
  SHOPPING: 'SHOPPING',
  OTHER: 'OTHER',
} as const;

export type CreateDailyLogBodySocialInteractionsItemSatisfactionLevel = typeof CreateDailyLogBodySocialInteractionsItemSatisfactionLevel[keyof typeof CreateDailyLogBodySocialInteractionsItemSatisfactionLevel];


export const CreateDailyLogBodySocialInteractionsItemSatisfactionLevel = {
  VERY_DISSATISFIED: 'VERY_DISSATISFIED',
  DISSATISFIED: 'DISSATISFIED',
  NEUTRAL: 'NEUTRAL',
  SATISFIED: 'SATISFIED',
  VERY_SATISFIED: 'VERY_SATISFIED',
} as const;